# docs/swagger.yaml
basePath: /api
definitions:
  boardhttp.boardResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      owner_id:
        type: string
      title:
        type: string
      updated_at:
        type: string
    type: object
  boardhttp.createBoardRequest:
    properties:
      description:
        type: string
      title:
        type: string
    required:
    - title
    type: object
  boardhttp.createBoardResponse:
    properties:
      id:
        type: string
    type: object
  boardhttp.updateBoardRequest:
    properties:
      description:
        type: string
      title:
        type: string
    type: object
...
host: localhost:8003
info:
  contact: {}
  description: API for managing kanban boards
  title: Student Kanban API
  version: "1.0"
paths:
...
  /boards:
    get:
      description: Returns a list of all boards owned by the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/boardhttp.boardResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get all boards for the authenticated user
      tags:
      - boards
    post:
      consumes:
      - application/json
      description: Creates a new board for the authenticated user
      parameters:
      - description: Board data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/boardhttp.createBoardRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/boardhttp.createBoardResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new board
      tags:
      - boards
  /boards/{id}:
    delete:
      description: Deletes a board by its UUID
      parameters:
      - description: Board UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a board
      tags:
      - boards
    get:
      description: Returns a single board by its UUID
      parameters:
      - description: Board UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/boardhttp.boardResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get a board by ID
      tags:
      - boards
    put:
      consumes:
      - application/json
      description: Updates a board's title and/or description. At least one field
        must be provided
      parameters:
      - description: Board UUID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/boardhttp.updateBoardRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update a board
      tags:
      - boards
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"